+++ src/features/updater/index.ts
@@ -1,0 +1,9 @@
+export { UpdaterBoot } from './UpdaterBoot'
+export {
+  defaultUpdaterDeps,
+  resetUpdaterDeps,
+  setUpdaterDeps,
+  useUpdaterStore,
+  type UpdaterDeps,
+  type UpdaterStatus,
+} from './updaterStore'